Time: 0 ms (100%), Space: 118.2 MB (82.7%) - LeetHub

+++ 2640-maximum-number-of-integers-to-choose-from-a-range-i/2640-maximum-number-of-integers-to-choose-from-a-range-i.cpp
@@ -110,8 +110,8 @@ class Solution {
              vis[val] = 1;//to ignore
          }
          int sum = 0;
-         int counter = 0;
-         for (int i = 1; i <= n; i++)
+         short counter = 0;
+         for (short i = 1; i <= n; i++)
          {
              if (vis[i] == 0)
              {
